feat: added Response entity status to the Client Controller

+++ src/main/java/com/example/CoffeeShop/controller/ClienteController.java
@@ -1,7 +1,9 @@
 package com.example.CoffeeShop.controller;
 
+import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -31,22 +33,36 @@ public class ClienteController {
   // Post the Client
   @CrossOrigin(origins = "*", allowedHeaders = "*")
   @PostMapping("addUser/client")
-  public void saveCliente(@RequestBody ClienteRequestDTO data) {
-    Cliente clienteData = new Cliente(data);
-    repository.save(clienteData);
-    return;
+  public ResponseEntity<Cliente> saveCliente(@RequestBody ClienteRequestDTO data) {
+    try {
+      Cliente clienteData = new Cliente(data);
+      repository.save(clienteData);
+      return new ResponseEntity<Cliente>(clienteData, HttpStatus.CREATED);
+    }
+
+    catch (Exception e) {
+      e.printStackTrace();
+      return new ResponseEntity<Cliente>(HttpStatus.BAD_REQUEST);
+    }
 
   }
 
   // Get all clients
   @CrossOrigin(origins = "*", allowedHeaders = "*")
   @GetMapping("getAllClients")
-  public List<ClienteResponseDTO> getAll() {
+  public ResponseEntity<List<ClienteResponseDTO>> getAll() {
+    try {
+      List<ClienteResponseDTO> clienteList = repository.findAll().stream().map(ClienteResponseDTO::new)
+          .toList();
 
-    List<ClienteResponseDTO> clienteList = repository.findAll().stream().map(ClienteResponseDTO::new)
-        .toList();
+      return new ResponseEntity<List<ClienteResponseDTO>>(clienteList, HttpStatus.FOUND);
+    }
 
-    return clienteList;
+    catch (Exception e) {
+      e.printStackTrace();
+      return new ResponseEntity<List<ClienteResponseDTO>>(HttpStatus.BAD_REQUEST);
+    }
+
   }
 
   // Get the client by Id
@@ -64,20 +80,28 @@ public class ClienteController {
   // Delete the Client by Id
   @CrossOrigin(origins = "*", allowedHeaders = "*")
   @DeleteMapping("deleteClient/{id_cliente}")
-  private void deleteCliente(@PathVariable Integer id_cliente) {
+  private ResponseEntity<String> deleteCliente(@PathVariable Integer id_cliente) {
     try {
-      repository.deleteById(id_cliente);
+      boolean deleted = repository.existsById(id_cliente);
+
+      if (deleted) {
+        repository.deleteById(id_cliente);
+        return ResponseEntity.ok("Resource deleted successfully");
+      } else {
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Resource not found");
+      }
     }
 
     catch (Exception e) {
       e.printStackTrace();
+      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Something went wrong");
     }
   }
 
   // Update the Client by Id
   @CrossOrigin(origins = "*", allowedHeaders = "*")
   @PutMapping("/updateClient/{id_cliente}")
-  private void updateClienteById(@PathVariable Integer id_cliente,
+  private ResponseEntity<Cliente> updateClienteById(@PathVariable Integer id_cliente,
       @RequestBody Cliente newFornecedorData) {
 
     try {
@@ -88,12 +112,17 @@ public class ClienteController {
         updateFornecedorData.setNm_cliente(newFornecedorData.getNm_cliente());
 
         repository.save(updateFornecedorData);
-        return;
+        return new ResponseEntity<Cliente>(updateFornecedorData, HttpStatus.OK);
+      }
+
+      else {
+        return new ResponseEntity<Cliente>(HttpStatus.BAD_REQUEST);
       }
     }
 
     catch (Exception e) {
       e.printStackTrace();
+      return new ResponseEntity<Cliente>(HttpStatus.NOT_MODIFIED);
     }
   }
 }
